refactor: move state_access.rs to reactive_state_access & updated imports
