+++ src/observable.rs
@@ -1,4 +1,4 @@
-use crate::state_access::{CloneState, StateAccess};
+use crate::reactive_state_access::state_access::{CloneState, StateAccess};
 
 use std::cell::RefCell;
 
@@ -19,59 +19,4 @@ where
     where
         T: Clone + 'static;
     fn observe_with<F: FnOnce(&T) -> R, R>(&self, func: F) -> R;
-}
-
-impl<T> Observable<T> for StateAccess<T>
-where
-    T: 'static,
-{
-    fn observe(&self) -> T
-    where
-        T: Clone,
-    {
-        let id = crate::store::StorageKey::TopoKey(self.id);
-        let context = illicit::get::<RefCell<ReactiveContext>>().expect(
-            "No #[reaction] context found, are you sure you are in one? I.e. does the current \
-             function have a #[reaction] tag?",
-        );
-        context.borrow_mut().reactive_state_accessors.push(id);
-
-        STORE.with(|store_refcell| {
-            store_refcell
-                .borrow_mut()
-                .add_dependency(&id, &context.borrow().key);
-        });
-
-        clone_reactive_state_with_id::<T>(id).unwrap()
-    }
-
-    #[topo::nested]
-    fn observe_update(&self) -> (Option<T>, T)
-    where
-        T: 'static + Clone,
-    {
-        let previous_value_access = crate::hooks_state_functions::use_state(|| None);
-        let opt_previous_value = previous_value_access.get();
-        let new_value = self.get();
-        previous_value_access.set(Some(new_value.clone()));
-        (opt_previous_value, new_value)
-    }
-
-    // <T: 'static, F: FnOnce(&T) -> R, R>(id: StorageKey, func: F) -> R {
-    fn observe_with<F: FnOnce(&T) -> R, R>(&self, func: F) -> R {
-        let id = crate::store::StorageKey::TopoKey(self.id);
-        if let Ok(context) = illicit::get::<RefCell<ReactiveContext>>() {
-            context
-                .borrow_mut()
-                .reactive_state_accessors
-                .push(id.clone());
-
-            STORE.with(|store_refcell| {
-                store_refcell
-                    .borrow_mut()
-                    .add_dependency(&id, &context.borrow().key);
-            });
-        }
-        read_reactive_state_with_id(id, func)
-    }
 }
