+++ src/reactive_state_access/state_access.rs
@@ -1,4 +1,7 @@
-use crate::{hooks_state_functions::*, store::TopoKey};
+use crate::{
+    clone_reactive_state_with_id, hooks_state_functions::*, read_reactive_state_with_id,
+    store::TopoKey, Observable, ReactiveContext,
+};
 use std::marker::PhantomData;
 
 ///  Accessor struct that provides access to getting and setting the
@@ -122,6 +125,8 @@ where
     }
 }
 
+use crate::reactive_state_functions::STORE;
+use std::cell::RefCell;
 use std::ops::{Add, Div, Mul, Sub};
 
 impl<T> Add for StateAccess<T>
@@ -165,5 +170,60 @@ where
 
     fn sub(self, other: Self) -> Self::Output {
         self.get_with(|s| other.get_with(|o| *o - *s))
+    }
+}
+
+impl<T> Observable<T> for StateAccess<T>
+where
+    T: 'static,
+{
+    fn observe(&self) -> T
+    where
+        T: Clone,
+    {
+        let id = crate::store::StorageKey::TopoKey(self.id);
+        let context = illicit::get::<RefCell<ReactiveContext>>().expect(
+            "No #[reaction] context found, are you sure you are in one? I.e. does the current \
+             function have a #[reaction] tag?",
+        );
+        context.borrow_mut().reactive_state_accessors.push(id);
+
+        STORE.with(|store_refcell| {
+            store_refcell
+                .borrow_mut()
+                .add_dependency(&id, &context.borrow().key);
+        });
+
+        clone_reactive_state_with_id::<T>(id).unwrap()
+    }
+
+    #[topo::nested]
+    fn observe_update(&self) -> (Option<T>, T)
+    where
+        T: 'static + Clone,
+    {
+        let previous_value_access = crate::hooks_state_functions::use_state(|| None);
+        let opt_previous_value = previous_value_access.get();
+        let new_value = self.get();
+        previous_value_access.set(Some(new_value.clone()));
+        (opt_previous_value, new_value)
+    }
+
+    // <T: 'static, F: FnOnce(&T) -> R, R>(id: StorageKey, func: F) -> R {
+    fn observe_with<F: FnOnce(&T) -> R, R>(&self, func: F) -> R {
+        let id = crate::store::StorageKey::TopoKey(self.id);
+        if let Ok(context) = illicit::get::<RefCell<ReactiveContext>>() {
+            context
+                .borrow_mut()
+                .reactive_state_accessors
+                .push(id.clone());
+
+            STORE.with(|store_refcell| {
+                store_refcell
+                    .borrow_mut()
+                    .add_dependency(&id, &context.borrow().key);
+            });
+        }
+        read_reactive_state_with_id(id, func)
     }
 }
